refactor(subway) 테스트 픽스쳐 클래스 분리하여 상속받도록 변경
