+++ src/test/java/nextstep/subway/domain/SectionsTest.java
@@ -4,59 +4,16 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import nextstep.subway.exception.ImpossibleDeleteException;
 import nextstep.subway.exception.NotFoundException;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class SectionsTest {
-    Station 강남역;
-    Station 양재역;
-    Station 판교역;
-    Station 양재시민의숲역;
-    Station 미정역;
-    Line 신분당선;
-    Section section1;
-    Section section2;
-    List<Section> sections = new ArrayList<>();
-    Sections sections1;
-
-    @BeforeEach
-    void setUp() {
-        강남역 = Station.builder("강남역")
-                .id(1L)
-                .build();
-        양재역 = Station.builder("양재역")
-                .id(2L)
-                .build();
-        판교역 = Station.builder("판교역")
-                .id(3L)
-                .build();
-        양재시민의숲역 = Station.builder("양재시민의숲역")
-                .id(4L)
-                .build();
-        미정역 = Station.builder("미정역")
-                .id(5L)
-                .build();
-        신분당선 = Line.builder("신분당선", "bg-red-600")
-                .id(1L)
-                .build();
-        section1 = Section.builder(강남역, 양재역, Distance.valueOf(10))
-                .build();
-        section2 = Section.builder(양재역, 양재시민의숲역, Distance.valueOf(10))
-                .build();
-        section1.addLine(신분당선);
-        section2.addLine(신분당선);
-        sections.add(section1);
-        sections.add(section2);
-        sections1 = Sections.valueOf(this.sections);
-    }
+class SectionsTest extends SectionsTestFixture {
 
     @DisplayName("구간들 사이에 새로운 구간 등록 테스트 (상행역 일치)")
     @Test
